requirements.txt
requests==2.31.0
urllib3==1.26.7
python-dotenv==1.0.0
websockets==12.0
pyTelegramBotAPI==4.14.0
telebot==0.0.5